# docs/swagger.yaml
basePath: /
definitions:
  controllers.ApproveRequest:
    properties:
      condition:
        type: string
    type: object
...
  models.Category:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.Donation:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      pickup_address:
        type: string
      pickup_date:
        type: string
      quantity:
        type: integer
      status:
        type: string
      tool_name:
        type: string
      user_detail:
        $ref: '#/definitions/models.UserResponse'
      user_id:
        type: string
    type: object
  models.InventoryLog:
    properties:
      action:
        type: string
      id:
        type: string
      note:
        type: string
      quantity:
        type: integer
      timestamp:
        type: string
      tool_id:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
host: api.medislink.web.id
info:
  contact: {}
  description: Medislink Backend API Documentation
  title: Medislink API
  version: "1.0"
paths:
...
  /api/categories/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create category
      tags:
      - Categories
  /api/donations/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all donations (admin)
      tags:
      - Donations
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Tool Name
        in: formData
        name: tool_name
        required: true
        type: string
      - description: Category
        in: formData
        name: category
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Pickup Address
        in: formData
        name: pickup_address
        required: true
        type: string
      - description: Pickup Date
        in: formData
        name: pickup_date
        required: true
        type: string
      - description: Quantity
        in: formData
        name: quantity
        required: true
        type: integer
      - description: Image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create donation
      tags:
      - Donations
  /api/donations/{id}/approve:
    put:
      consumes:
      - application/json
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: string
      - description: QC Condition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve donation and add stock
      tags:
      - Donations
  /api/donations/{id}/receive:
    put:
      parameters:
      - description: Donation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark donation as received
      tags:
      - Donations
  /api/donations/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Donation'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user donations
      tags:
      - Donations
  /api/inventory-logs/:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.InventoryLog'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all inventory logs
      tags:
      - InventoryLogs
    post:
      consumes:
      - application/json
      parameters:
      - description: Inventory Log Data
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/models.InventoryLog'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create inventory log
      tags:
      - InventoryLogs
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"